Move SmartFormatter to utils module
